new method for creating order

+++ src/main/java/org/example/logisticapplication/domain/Cargo/CargoInfoDto.java
@@ -1,11 +1,10 @@
 package org.example.logisticapplication.domain.Cargo;
 
-import java.math.BigDecimal;
 
 public record CargoInfoDto(
         String number,
         String name,
-        BigDecimal weightKg,
+        Double weightKg,
         String status
 ) {
 }
